+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * unreleased
   - wrap DDL in transaction so partial changes don't apply on sqlite
+  - support creating, but not yet diffing, enums
 * 0.1.2
   - internal: looser yaml dep for compat with other packages
   - bugfix: don't drop columns on sqlite (because sqlite doesn't drop columns)